refactor: split graph ql queries
